backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
scikit-learn==1.5.1
pandas
joblib
python-multipart
python-dotenv
pytest
httpx